basic setup

+++ src/main/java/pmrc/at/postgres/persistence/UserRepository.java
@@ -1,9 +1,9 @@
 package pmrc.at.postgres.persistence;
 
-import org.apache.catalina.User;
+import pmrc.at.postgres.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
